requirements.txt
# Discord bot framework
discord.py>=2.3.0

# Environment variable management
python-dotenv>=1.0.0

# OpenAI API client
openai>=0.27.0

# Anthropic Claude API client
anthropic>=0.2.0

# Google Gemini AI API client
google-generativeai>=0.1.0

# Requests library for making HTTP requests
requests>=2.28.0

# JSON handling
ujson>=5.4.0

# SQLAlchemy for database interactions
sqlalchemy>=1.4.0

# Alembic for database migrations
alembic>=1.7.0

# Pytest for testing
pytest>=7.1.0

# Black for code formatting
black>=22.3.0

# Flask for web server
flask>=2.0.0

# Aiohttp for asynchronous HTTP requests
aiohttp>=3.8.0

# Aiosqlite for asynchronous SQLite interactions
aiosqlite>=0.17.0

# Flask for web server
flask>=2.0.0

# Aiohttp for asynchronous HTTP requests
aiohttp>=3.8.0

# Aiosqlite for asynchronous SQLite interactions
aiosqlite>=0.17.0

# Logging configuration
loguru>=0.6.0

transformers>=4.0.0
torch>=1.7.0
praw
pytesseract